Re-organized trec-cast
- Moved generic classes out of trec_cast.py
- Fixed some bugs and problems
